(refactor): Namespace restructuring
