+++ test/ContentFileScanner.test.ts
@@ -5,15 +5,15 @@ import { withDir } from 'tmp-promise';
 
 import { ContentFileScanner } from '../src/ContentFileScanner';
 import ContentManager from '../src/ContentManager';
+import EditorConfig from '../src/implementation/EditorConfig';
+import FileContentStorage from '../src/implementation/fs/FileContentStorage';
+import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
 import { ContentId, IUser } from '../src/types';
 
-import EditorConfig from '../examples/EditorConfig';
 import User from '../examples/User';
-import FileContentStorage from '../src/implementation/fs/FileContentStorage';
-import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 
 import { getContentDetails } from './ContentScanner.test';
