+++ test/ContentScanner.test.ts
@@ -4,15 +4,15 @@ import { withDir } from 'tmp-promise';
 
 import ContentManager from '../src/ContentManager';
 import { ContentScanner } from '../src/ContentScanner';
+import EditorConfig from '../src/implementation/EditorConfig';
+import FileContentStorage from '../src/implementation/fs/FileContentStorage';
+import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
 import { ContentId, ILibraryName, IUser } from '../src/types';
 
-import EditorConfig from '../examples/EditorConfig';
 import User from '../examples/User';
-import FileContentStorage from '../src/implementation/fs/FileContentStorage';
-import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 
 async function createContentScanner(
     file: string,
