+++ test/DependencyGetter.test.ts
@@ -1,10 +1,9 @@
 import path from 'path';
 
 import DependencyGetter from '../src/DependencyGetter';
+import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import LibraryName from '../src/LibraryName';
-
-import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 
 describe('basic file library manager functionality', () => {
     it('determines dependencies of libraries', async () => {
