+++ test/PackageImporter.test.ts
@@ -5,15 +5,14 @@ import { BufferWritableMock } from 'stream-mock';
 import { withDir } from 'tmp-promise';
 
 import ContentManager from '../src/ContentManager';
-
+import EditorConfig from '../src/implementation/EditorConfig';
+import FileContentStorage from '../src/implementation/fs/FileContentStorage';
+import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
 
-import EditorConfig from '../examples/EditorConfig';
 import User from '../examples/User';
-import FileContentStorage from '../src/implementation/fs/FileContentStorage';
-import FileLibraryStorage from '../src/implementation/fs/FileLibraryStorage';
 
 describe('package importer', () => {
     it('installs libraries', async () => {
